Update keep-presence.py
The mouse movement has been changed from a linear movement that keeps going to a back and forth movement. It will move and then move right back to it's original position, keeping it from traveling across the screen.

There has also been a timeout parameter added that will allow you to specify a length of time for the program to run. So you can have it run for a specific duration and then stop itself and allow the computer to continue with it's normal power save functions instead of being kept awake.